requirements.txt
astroid==2.2.5
certifi==2021.5.30
colorama==0.4.1
cycler==0.10.0
fonttools==4.25.0
isort==4.3.21
joblib
kiwisolver 
lazy-object-proxy==1.4.1
matplotlib 
mccabe==0.6.1
mkl-fft==1.3.0
mkl-random==1.0.4
mkl-service==2.3.0
munkres==1.1.4
numpy 
olefile==0.46
Pillow 
pyparsing 
python-dateutil 
scikit-learn 
scipy 
six 
threadpoolctl 
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0
tornado
tqdm
typed-ast==1.4.0
typing-extensions 
wincertstore==0.2
wrapt==1.11.2
